Rename project to NucFlag.
