+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools", "setuptools-scm"]
+requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
@@ -10,21 +10,21 @@ dependencies = {file = ["requirements.txt"]}
 optional-dependencies = {dev = { file = ["requirements-dev.txt"] }}
 
 [tool.setuptools.packages.find]
-include = ["nucfreq*"]
+include = ["nucflag*"]
 exclude = ["docs*", "test*"]
 
 [project]
-name = "nucfreq"
+name = "nucflag"
 version = "0.0.2"
 authors = [
     {name = "Keith Oshima", email = "oshimak@pennmedicine.upenn.edu"},
     {name = "Mitchell R. Vollger", email = "mrvollger@gmail.com"},
 ]
-description = "Nucfreq misassembly identifier."
+description = "NucFlag misassembly identifier."
 readme = "README.md"
 requires-python = ">=3.12"
 license = {text = "MIT License"}
 dynamic = ["dependencies"]
 
 [project.scripts]
-nucfreq = "nucfreq.main:main"
+nucflag = "nucflag.main:main"
