+++ test/workflow/env.yaml
@@ -4,5 +4,5 @@ dependencies:
   - python=3.12.0
   - pip
   - pip:
-      - ../../dist/nucfreq-0.0.2-py3-none-any.whl
-      # - git+https://github.com/logsdon-lab/NucFreq.git
+      - ../../dist/nucflag-0.0.2-py3-none-any.whl
+      # - git+https://github.com/logsdon-lab/NucFlag.git
